Refactoring: moved TTCrystal to a separate file
